create electrum module
